Fixed bugs in the engine while expanding on the game and its levels.

+++ Project Aela/3D/Particles/PointParticleEmitter.h
@@ -30,6 +30,6 @@ namespace Aela {
 			// This class requires a pointer to the camera so that it can properly sort particles.
 			Camera3D* camera;
 
-			void setupParticlePositioning(unsigned int whichParticle, unsigned int numberOfParticles);
+			void setupParticlePositioning(size_t whichParticle, size_t numberOfParticles);
 	};
 }
